tensorflow: lite: label_image.cc
Modify part of the tensorflow lite source code to support the penglai system.
Add dependencies on methods in the penglai system when compiling the riscv version of tensorflow lite.
Add the penglai_demo folder under tensorflow lite, containing a tensorflow lite application that can run on the penglai system and can be compiled via make.

Signed-off-by: Moonquakes <467946553@qq.com>